+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Copy to Clipboard button to TabFileImage toolbar
 - ScreenCapture Border, Icon, and Memory Improvements
 - TextEditor Search Panel Theming
+- Added FileTypeDetector to allow probing files to determine their type
 - Toolbar Copy button to TabTextFile
 - CHANGELOG.md to track version history
 - CONTRIBUTING.md for contributor guidelines